feat - separte getFormData from ImgUpload
